feat(286): fix filters panel styling for mobile screens
