+++ src/theme/globalStyles.js
@@ -18,17 +18,19 @@ const globalStyles = globalCss({
     '-moz-osx-font-smoothing': 'grayscale',
 
     // Date Picker styles
-    // Had to override them here b/c the actual Date Picker element gets 
-    // appended to the body (and thus you can't dump style overrides in a 
+    // Had to override them here b/c the actual Date Picker element gets
+    // appended to the body (and thus you can't dump style overrides in a
     // wrapper element)
 
     '.DateInput_input': {
       color: '$textDark',
       fontFamily: '$mono',
       fontWeight: '$2',
+      fontSize: '$4',
     },
     '.DateInput__small': {
-      width: '109px',
+      width: '115px',
+      fontSize: '$4',
     },
     '.DateInput_input__focused': {
       borderBottomColor: '$blue600',
